requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.7.0
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
ply==3.11
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
PyQt5==5.9
PyQt5-sip==12.7.0
pyqt5-tools==5.13.0.1.5
PyQt5Designer==5.13.1
python-dotenv==0.10.3
pytz==2020.1
selenium==3.141.0
sip==4.19.8
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.31
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.2.0
